airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pandas
pyarrow
azure-storage-blob
azure-identity
python-dotenv
snowflake-connector-python
snowflake-connector-python[pandas]
sqlalchemy
snowflake-sqlalchemy